feat: simplify ext rpc protocol
