+++ packages/extension/__mocks__/initRPCProtocol.ts
@@ -1,5 +1,6 @@
 import { RPCServiceCenter, initRPCService } from '@opensumi/ide-connection';
 import { RPCProtocol } from '@opensumi/ide-connection/lib/common/ext-rpc-protocol';
+import { Emitter } from '@opensumi/ide-core-common';
 
 export async function initMockRPCProtocol(client): Promise<RPCProtocol> {
   const extCenter = new RPCServiceCenter();
@@ -15,4 +16,25 @@ export async function initMockRPCProtocol(client): Promise<RPCProtocol> {
   });
 
   return extProtocol;
+}
+
+export function createMockPairRPCProtocol() {
+  const emitterA = new Emitter<any>();
+  const emitterB = new Emitter<any>();
+
+  const mockClientA = {
+    send: (msg) => emitterB.fire(msg),
+    onMessage: emitterA.event,
+  };
+  const mockClientB = {
+    send: (msg) => emitterA.fire(msg),
+    onMessage: emitterB.event,
+  };
+
+  const rpcProtocolExt = new RPCProtocol(mockClientA);
+  const rpcProtocolMain = new RPCProtocol(mockClientB);
+  return {
+    rpcProtocolExt,
+    rpcProtocolMain,
+  };
 }
